Generate random number before pressing a button
Added a Start button that calls a function, that generate a random number (a winning button).
Now the random number is generated before pressing one button.
Also the random number, for 1'st play round is generated even before pressing the Start button. This allows the user to play the first round without boring him to press the Start button. But after the 1'st round he will have to hit the Start button to play another one.